requirements.txt
requests
beautifulsoup4
jira
python-dotenv
selenium
webdriver-manager
httpx
selenium-wire